tickethub_back/users/urls.py
"""Users urls."""

# Django
from django.urls import path, include

# Django REST Framework
from rest_framework.routers import DefaultRouter
from rest_framework_simplejwt.views import TokenRefreshView

# Views
from tickethub_back.users import views

router = DefaultRouter()
router.register(r'users', views.UserViewSet, basename='user')

urlpatterns = [
    path('', include(router.urls)),
    path('users/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
]