requirements/base.txt
asgiref==3.7.2

# django
Django==3.2.12
django-cors-headers==3.14.0
django-extensions==3.2.3
django-filter==23.3
django-environ==0.11.2
django-password-validators==1.7.1

# rest frame work
djangorestframework==3.14.0
djangorestframework-filters==1.0.0.dev2
djangorestframework-simplejwt==5.3.0


Pillow==9.5.0
pluggy==1.2.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
termcolor==2.3.0
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1

# env
python-dotenv==0.21.1


